requirements.txt
# Core LLM APIs
tavily-python>=0.3.0
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.21.0

# Configuration Management
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0.0

# Database
SQLAlchemy>=2.0.0

# File Handling
Pillow>=10.0.0
PyPDF2>=3.0.0
python-magic>=0.4.27

# Token Counting
tiktoken>=0.5.0

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3

# Frontend (Streamlit)
streamlit>=1.30.0
streamlit-chat>=0.1.1

# Testing (Development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Code Quality (Development)
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0

